Tela de edição de guarnição funcionando

+++ tiro-guerra/wwwroot/js/ModalControleUpdateGuarnicao.js
@@ -6,7 +6,7 @@ var id_dia;
 
 
 //modal
-
+//exibe o modal na tela
 function showModal(idModal)
 {
     id_dia = idModal;   
@@ -14,13 +14,14 @@ function showModal(idModal)
     element.classList.add("show-modal");
 }
 
+//esconde o modal
 function disposeModal()
 {
     var element = document.getElementById("modal");
     element.classList.remove("show-modal");
 }
 
-
+//deleta um sentinela da lista de sentinelas
 function deleteItem(atirador)
 {
     let listaAtiradores = document.querySelector('.myList');
@@ -60,11 +61,11 @@ function deleteItem(atirador)
 
      }
 
-     console.log('Quantidade de atiradores '+atiradores.length);
+     //console.log('Quantidade de atiradores '+atiradores.length);
 
 }
 
-//adiciona 1 sentinela a lista
+//adiciona 1 sentinela a lista de sentinelas
 document.querySelector('input[value=Adicionar]').addEventListener('click',() =>
 {
    
@@ -74,38 +75,38 @@ document.querySelector('input[value=Adicionar]').addEventListener('click',() =>
 
     if(atiradores.length < 6)
     {
-
-        atiradores.push(value)
-    
-        atiradoresLista.push({
-            id:valueId,
-            nome: value,
-            funcao: 'Sentinela'
+        let atiradorEstaNaLista = false;
+        atiradores.forEach(function (item, indice)
+        {
+            if(value == item)
+            {
+                atiradorEstaNaLista = true;
+            }
         });
 
-        let listaAtiradores = document.querySelector('.myList');
-        listaAtiradores.innerHTML += `
-        <div class="linha-lista">
-            <div class="list-space">
-                <li class='item-lista' id='`+value+`'> `+value+`</li>
-            </div>
-
-            <div class="list-space">
-                <span  class='span-item-lista' onClick="deleteItem('`+value+`')"> X </span>
-            </div> 
-        </div>`
-        
-         
-       //  <li id="Amaro">Teste  <span onclick="deleteItem('Amaro')">X</span> </li>  
-
-       /*
-          <div class="list-space">
-                                        <li class="item-lista" id="Teste">Teste   </li>
-                                    </div>
-                                    <div class="list-space">
-                                         <span class="span-item-lista" onclick="deleteItem('Teste')">X</span> 
-                                    </div> 
-       */
+        if(atiradorEstaNaLista == false)
+        {
+            atiradores.push(value);
+    
+            atiradoresLista.push({
+                id:valueId,
+                nome: value,
+                funcao: 'Sentinela'
+            });
+    
+            let listaAtiradores = document.querySelector('.myList');
+            listaAtiradores.innerHTML += `
+            <div class="linha-lista">
+                <div class="list-space">
+                    <li class='item-lista' id='`+value+`'> `+value+`</li>
+                </div>
+    
+                <div class="list-space">
+                    <span  class='span-item-lista' onClick="deleteItem('`+value+`')"> X </span>
+                </div> 
+            </div>`
+        }
+   
     }
 
     select = "";
@@ -129,9 +130,7 @@ document.querySelector('input[value=Limpar]').addEventListener('click',() =>
 //Recebe as informações dos sentinelas e dos outros membros da guarda
 document.querySelector('input[class=save]').addEventListener('click',() =>
 {
-
-    
-
+    //Leitura dos valores descritos no select
     var select = document.getElementById('fiscal');
     var FiscalId = select.options[select.selectedIndex].value;
     var fiscalName = select.options[select.selectedIndex].text;
@@ -144,27 +143,30 @@ document.querySelector('input[class=save]').addEventListener('click',() =>
     var CaboID = select.options[select.selectedIndex].value;
     var CaboName = select.options[select.selectedIndex].text;
 
-    console.log(atiradoresLista);
-      
+    //console.log(atiradoresLista);
+    //atualiza o valor do dia para bater com o indice 0 do array
     var index = id_dia -1;
 
+    //leitura dos valores do id, nome e id da guarnição do fiscal que sera alterado
     let linhaFiscaisSemanaId = document.querySelectorAll('#linha-fiscal td .idFiscal');
     let linhaFiscaisSemanaNome = document.querySelectorAll('#linha-fiscal td .nome');
     let linhaFiscaisSemanaIdguarnicao = document.querySelectorAll('#linha-fiscal td .idGuarnicao');
 
-
+    //leitura dos valores do id, nome e id da guarnição do comandante que sera alterado
     let linhaComandantesSemanaId = document.querySelectorAll('#linha-comandante td .idComandante');
     let linhaComandantesSemanaNome = document.querySelectorAll('#linha-comandante td .nome');
     let linhaComandantesSemanaIdGuarda = document.querySelectorAll('#linha-comandante td .idGuarda');
 
-
+    //leitura dos valores do id, nome e id da guarnição do cabo que sera alterado
     let linhaCabosSemanaId = document.querySelectorAll('#linha-cabo td .IdCabos');
     let linhaCabosSemanaNome = document.querySelectorAll('#linha-cabo td .nome');
     let linhaCabosSemanaIdGuarda = document.querySelectorAll('#linha-cabo td .idGuarda');
 
+    //remove o campo nome e id do atirador que sera alterado
     linhaFiscaisSemanaId[index].remove();
     linhaFiscaisSemanaNome[index].remove();
 
+    //Adiciona o novo item após o id da guarda
     linhaFiscaisSemanaIdguarnicao[index].insertAdjacentHTML("afterend",`
             <input class="idFiscal" name="idfiscal" type="hidden" value="`+FiscalId+`">
             <label class="nome" name="nome">`+fiscalName+`</label>
@@ -172,11 +174,6 @@ document.querySelector('input[class=save]').addEventListener('click',() =>
 
     linhaComandantesSemanaId[index].remove();
     linhaComandantesSemanaNome[index].remove();
-
-    console.log(linhaFiscaisSemanaIdguarnicao);
-    console.log(linhaComandantesSemanaIdGuarda);
-    console.log(linhaCabosSemanaIdGuarda);
-
     
     linhaComandantesSemanaIdGuarda[index].insertAdjacentHTML("afterend",`
     <input class="idComandante" name="idAtirador" type="hidden" value="`+ComandanteId+`">
@@ -190,47 +187,37 @@ document.querySelector('input[class=save]').addEventListener('click',() =>
     <input class="IdCabos" name="idAtirador" type="hidden" value="`+CaboID+`">
     <label class="nome" name="nome">`+CaboName+`</label>`);
 
-    /*
-    linhaFiscaisSemana[index].innerHTML =`
-    <input name="idfiscal" type="hidden" value="`+FiscalId+`"> 
-    <label name="nome">`+fiscalName+`</label> 
-    `;
-
-    linhaComandantesSemana[index].innerHTML = `
-    <input name="id" type="hidden" value="`+ComandanteId+`"> 
-    <input name="Funcao" type="hidden" value="Comandante">
-    <label name="nome">`+ComandanteName+`</label> 
-    `;
-
-    linhaCabosSemana[index].innerHTML = `
-    <input name="id" type="hidden" value="`+CaboID+`"> 
-    <input name="Funcao" type="hidden" value="Cabo">
-    <label name="nome">`+CaboName+`</label> 
-    `
-*/
     let classe;
-    let linhaSentinelas;
     var contador = 1;
 
+    let linhaSentinelaSemanaId;
+    let linhaSentinelaSemanaNome;
+    let linhaSentinelaSemanaIdGuarda;
+
+    //percorre a lista de sentinelas para atualizar a tela
     atiradoresLista.forEach(function (item, indice, array)
      {
-        let linhaIdAtirador;
-        let linhaNomeAtirador;
-        let linhaIdGuarda;
 
-        // Alterar aqui para ficar semelhante aos de cima
-        classe ='#linha-sentinela0'+contador+' td .idAtirador';
+        //Recupera a linha que sera alterada
+        classe ='#linha-sentinela0'+contador+' td';
         contador+=1;
 
-        linhaSentinelas = document.querySelectorAll(classe);
-        linhaSentinelas[index].innerHTML = `
-        <input name="id" type="hidden" value="`+item.id+`"> 
-        <input name="Funcao" type="hidden" value="Sentinela">
-        <label name="nome">`+item.nome+`</label> 
-        `;
+        //recebe os valores atuais
+        linhaSentinelaSemanaId = document.querySelectorAll(classe+" .idAtirador");
+        linhaSentinelaSemanaNome = document.querySelectorAll(classe+" .nome");
+        linhaSentinelaSemanaIdGuarda = document.querySelectorAll(classe+" .idGuarda");
+
+        //remove os valores de nome e id
+        linhaSentinelaSemanaId[index].remove();
+        linhaSentinelaSemanaNome[index].remove();
+
+        //insere ao final os novos valores de nome e Id
+        linhaSentinelaSemanaIdGuarda[index].insertAdjacentHTML("afterend",
+        `<input class="idAtirador" name="idAtirador" type="hidden" value="`+item.id+`">
+        <label class="nome" name="nome">`+item.nome+`</label>`);
     });
 
-    
+    //Limpa a lista do modal
     let listaAtiradores = document.querySelector('.myList');
     listaAtiradores.innerHTML = ""
